requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
biopython==1.85
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dm-tree==0.1.9
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fair-esm==2.0.0
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
ml_collections==1.1.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rdkit==2022.9.5
rdkit-pypi==2022.9.5
requests==2.32.3
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
torchdrug==0.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
